requirements.txt
asgiref==3.6.0
certifi @ file:///C:/ci_311/certifi_1676421629969/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-pagination==1.0.7
django-tinymce==3.6.1
django-tinymce4==3.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
mssql-django==1.3
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==1.7.1
pyodbc==4.0.39
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sentry-sdk==1.18.0
smmap==5.0.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
wincertstore==0.2
